Only Copy Compartment self.config when Necessary
When building a compartment for wcEcoli, we discovered that the wcEcoli
agent process cannot be copied with copy.deepcopy because it cannot be
serialized. As a temporary fix, this commit avoids copying a
compartment's instance field configuration when, as is the case for
wcEcoli, we don't pass a separate config to generate().