Use UTC Timestamp for Default Experiment ID
We used to use a UUID for the default experiment ID, but this
communicated no useful information. Using a timestamp adds meaning to
the IDs, making it easier to identify an experiment by its ID. This has
the added benefit of meaning that any outputs named with the experiment
ID will get sorted in sequential order.

We already emit a timezone-adjusted timestamp from each experiment, but
we use a UTC timestamp here to make IDs more reliably unique. With UTC,
experiment IDs will always be sequential, even if the user changes
timezones.