Add Tagged Molecule Analysis Plotter
When molecules have been tagged by the simulation and stored in a store
in an agent's boundary store, we can create plots to mimic the images we
might get if the molecules were fluorescently tagged. For example, with
a molecule tagged CDPDIGLYSYN-MONOMER[i], we can create a CSV file
tags.csv containing:
        bulk_molecules_report,CDPDIGLYSYN-MONOMER[i]
Then if we run python -m vivarium.analysis.analyze -fg tags.csv exp_id,
we'll get a plot based on the molecule's concentration (counts read from
the bulk_molecules_report store).